feat: Access Token

+++ src/app/api/users/route.ts
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from '@/lib/firebase/route';
 import { NextRequest, NextResponse } from 'next/server';
+import jwt from 'jsonwebtoken';
 
 export async function GET(req: NextRequest) {
   const users = await retrieveData('users');
@@ -13,12 +14,28 @@ export async function GET(req: NextRequest) {
 
 export async function PUT(req: NextRequest) {
   const { id, data } = await req.json();
-  const result = await updateData('users', id, data);
-  return NextResponse.json({ status: result.status, message: result.message }, { status: result.statusCode });
+  const reqHeader = new Headers(req.headers);
+  const token = reqHeader.get('Authorization')?.split(' ')[1] || '';
+  return jwt.verify(token, process.env.NEXTAUTH_SECRET || '', async (err: any, decoded: any) => {
+    if (decoded && decoded.role === 'admin') {
+      const result = await updateData('users', id, data);
+      return NextResponse.json({ status: result.status, message: result.message }, { status: result.statusCode });
+    } else {
+      return NextResponse.json({ status: false, message: 'Access denied' }, { status: 403 });
+    }
+  });
 }
 
 export async function DELETE(req: NextRequest) {
   const { id } = await req.json();
-  const result = await deleteData('users', id);
-  return NextResponse.json({ status: result.status, message: result.message }, { status: result.statusCode });
+  const reqHeader = new Headers(req.headers);
+  const token = reqHeader.get('Authorization')?.split(' ')[1] || '';
+  return jwt.verify(token, process.env.NEXTAUTH_SECRET || '', async (err: any, decoded: any) => {
+    if (decoded && decoded.role === 'admin') {
+      const result = await deleteData('users', id);
+      return NextResponse.json({ status: result.status, message: result.message }, { status: result.statusCode });
+    } else {
+      return NextResponse.json({ status: false, message: 'Access denied' }, { status: 403 });
+    }
+  });
 }
